Backend/API_Wrappers/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
aiosqlite==0.21.0
neo4j==5.28.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
